I exported my function to a separate .js, and export/imported in necessary files
